pyproject.toml
[project]
name = "query-genie"
authors = [
  {name = "Michal Galuszka"},
]
version = "1.0.0"
description = "AI Chat Application with Strands Agents and PostgreSQL MCP Server"
requires-python = "==3.13.*"
dependencies = [
    "python-dotenv==1.2.1",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
]

[dependency-groups]
backend = [
    "fastapi==0.127.0",
    "uvicorn==0.40.0",
    "slowapi==0.1.9",
    "redis==7.1.0",
    "sse-starlette==3.0.4",
    "Jinja2==3.1.6",
    "httpx==0.28.1",
]
agent-service = [
    "boto3==1.42.15",
    "strands-agents[ollama,openai]==1.20.0",
    "strands-agents-tools==0.2.18",
    "mcp[cli]==1.25.0",
    "redis==7.1.0",
    "httpx==0.28.1",
]
mcp-server = [
    "mcp==1.25.0",
    "asyncpg==0.31.0",
    "sqlparse==0.5.5",
]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "rust-just>=1.45.0",
    "bandit>=1.9.2",
    "pip-audit>=2.10.0",
]
commit = ["commitizen>=4.10.1", "pre-commit>=4.5.1"]

# Tool configurations
[tool.bandit]
targets = ["Apps"]
severity = "all"
skips = ["B608"]  # SQL injection - false positives on validated identifiers
exclude_dirs = [
    "frontend/*",
    "*/static/*",
    "*/strands_sessions/*",
]
exclude = [
    "*.js",
    "*.jsx",
    "*.tsx",
    "*.ts",
    "*.css",
    "*.scss",
    "*.svg",
    "*.png",
    "*.jpeg",
    "*.jpg",
    "*.txt",
    "*.bat",
    "*.sh",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["misc", "untyped-decorator", "no-untyped-call", "arg-type"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests", "mcp_postgres/tests", "agent_service/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = ["__pycache__", ".git", ".venv"]
# Coverage options specified in just test tasks, not here
filterwarnings = [
    "ignore::DeprecationWarning:strands.experimental.hooks",
]
addopts = [
    "-v",
    "--tb=short",
]

[tool.coverage.run]
# Source is specified per-module in just test tasks using --cov= flags
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/__init__.py",
    # Entry points and infrastructure files
    "backend/src/app.py",
    "backend/src/api/middleware.py",
    "backend/src/api/rate_limit.py",
    "backend/src/logging_config.py",
    "mcp_postgres/src/server.py",
    "mcp_postgres/src/core/logging_config.py",
    "agent_service/src/main.py",
    "agent_service/src/utils/logging_config.py",
    "frontend/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 70
